+++ lib/processor/protobuf.go
@@ -55,7 +55,8 @@ Attempts to create a target protobuf message from a generic JSON structure.`,
 		FieldSpecs: docs.FieldSpecs{
 			docs.FieldCommon("operator", "The [operator](#operators) to execute").HasOptions("to_json", "from_json"),
 			docs.FieldCommon("message", "The fully qualified name of the protobuf message to convert to/from."),
-			docs.FieldCommon("import_path", "A path to a .proto file, or directory containing all .proto files required for parsing the target message. If left empty the current directory is used."),
+			docs.FieldCommon("import_paths", "A list of directories containing .proto files, including all definitions required for parsing the target message. If left empty the current directory is used. Each directory listed will be walked with all found .proto files imported."),
+			docs.FieldDeprecated("import_path"),
 			partsFieldSpec,
 		},
 		Examples: []docs.AnnotatedExample{
@@ -99,7 +100,7 @@ pipeline:
     - protobuf:
         operator: from_json
         message: testing.Person
-        import_path: testing/schema
+        import_paths: [ testing/schema ]
 `,
 			},
 			{
@@ -142,7 +143,7 @@ pipeline:
     - protobuf:
         operator: to_json
         message: testing.Person
-        import_path: testing/schema
+        import_paths: [ testing/schema ]
 `,
 			},
 		},
@@ -153,19 +154,21 @@ pipeline:
 
 // ProtobufConfig contains configuration fields for the Protobuf processor.
 type ProtobufConfig struct {
-	Parts      []int  `json:"parts" yaml:"parts"`
-	Operator   string `json:"operator" yaml:"operator"`
-	Message    string `json:"message" yaml:"message"`
-	ImportPath string `json:"import_path" yaml:"import_path"`
+	Parts       []int    `json:"parts" yaml:"parts"`
+	Operator    string   `json:"operator" yaml:"operator"`
+	Message     string   `json:"message" yaml:"message"`
+	ImportPaths []string `json:"import_paths" yaml:"import_paths"`
+	ImportPath  string   `json:"import_path" yaml:"import_path"`
 }
 
 // NewProtobufConfig returns a ProtobufConfig with default values.
 func NewProtobufConfig() ProtobufConfig {
 	return ProtobufConfig{
-		Parts:      []int{},
-		Operator:   "to_json",
-		Message:    "",
-		ImportPath: "",
+		Parts:       []int{},
+		Operator:    "to_json",
+		Message:     "",
+		ImportPaths: []string{},
+		ImportPath:  "",
 	}
 }
 
@@ -173,8 +176,8 @@ func NewProtobufConfig() ProtobufConfig {
 
 type protobufOperator func(part types.Part) error
 
-func newProtobufToJSONOperator(message, importPath string) (protobufOperator, error) {
-	m, err := loadDescriptor(message, importPath)
+func newProtobufToJSONOperator(message string, importPaths []string) (protobufOperator, error) {
+	m, err := loadDescriptor(message, importPaths)
 	if err != nil {
 		return nil, err
 	}
@@ -194,8 +197,8 @@ func newProtobufToJSONOperator(message, importPath string) (protobufOperator, er
 	}, nil
 }
 
-func newProtobufFromJSONOperator(message, importPath string) (protobufOperator, error) {
-	m, err := loadDescriptor(message, importPath)
+func newProtobufFromJSONOperator(message string, importPaths []string) (protobufOperator, error) {
+	m, err := loadDescriptor(message, importPaths)
 	if err != nil {
 		return nil, err
 	}
@@ -215,44 +218,45 @@ func newProtobufFromJSONOperator(message, importPath string) (protobufOperator, 
 	}, nil
 }
 
-func strToProtobufOperator(opStr, message, importPath string) (protobufOperator, error) {
+func strToProtobufOperator(opStr, message string, importPaths []string) (protobufOperator, error) {
 	switch opStr {
 	case "to_json":
-		return newProtobufToJSONOperator(message, importPath)
+		return newProtobufToJSONOperator(message, importPaths)
 	case "from_json":
-		return newProtobufFromJSONOperator(message, importPath)
+		return newProtobufFromJSONOperator(message, importPaths)
 	}
 	return nil, fmt.Errorf("operator not recognised: %v", opStr)
 }
 
-func loadDescriptor(message, importPath string) (*desc.MessageDescriptor, error) {
+func loadDescriptor(message string, importPaths []string) (*desc.MessageDescriptor, error) {
 	if len(message) == 0 {
 		return nil, errors.New("message field must not be empty")
 	}
 
 	var parser protoparse.Parser
-	if len(importPath) > 0 {
-		parser.ImportPaths = []string{importPath}
+	if len(importPaths) == 0 {
+		importPaths = []string{"."}
 	} else {
-		importPath = "."
+		parser.ImportPaths = importPaths
 	}
 
 	var files []string
-	err := filepath.Walk(importPath, func(path string, info os.FileInfo, err error) error {
-		if err != nil || info.IsDir() {
-			return err
-		}
-		if filepath.Ext(info.Name()) == ".proto" {
-			rPath, err := filepath.Rel(importPath, path)
-			if err != nil {
-				return fmt.Errorf("failed to get relative path: %v", err)
+	for _, importPath := range importPaths {
+		if err := filepath.Walk(importPath, func(path string, info os.FileInfo, ferr error) error {
+			if ferr != nil || info.IsDir() {
+				return ferr
 			}
-			files = append(files, rPath)
+			if filepath.Ext(info.Name()) == ".proto" {
+				rPath, ferr := filepath.Rel(importPath, path)
+				if ferr != nil {
+					return fmt.Errorf("failed to get relative path: %v", ferr)
+				}
+				files = append(files, rPath)
+			}
+			return nil
+		}); err != nil {
+			return nil, err
 		}
-		return nil
-	})
-	if err != nil {
-		return nil, err
 	}
 
 	fds, err := parser.ParseFiles(files...)
@@ -260,7 +264,7 @@ func loadDescriptor(message, importPath string) (*desc.MessageDescriptor, error)
 		return nil, fmt.Errorf("failed to parse .proto file: %v", err)
 	}
 	if len(fds) == 0 {
-		return nil, fmt.Errorf("no .proto files were found in the path '%v'", importPath)
+		return nil, fmt.Errorf("no .proto files were found in the paths '%v'", importPaths)
 	}
 
 	var msg *desc.MessageDescriptor
@@ -270,7 +274,7 @@ func loadDescriptor(message, importPath string) (*desc.MessageDescriptor, error)
 		}
 	}
 	if msg == nil {
-		err = fmt.Errorf("unable to find message '%v' definition within '%v'", message, importPath)
+		err = fmt.Errorf("unable to find message '%v' definition within '%v'", message, importPaths)
 	}
 	return msg, err
 }
@@ -308,8 +312,13 @@ func NewProtobuf(
 		mBatchSent: stats.GetCounter("batch.sent"),
 	}
 
+	importPaths := conf.Protobuf.ImportPaths
+	if len(conf.Protobuf.ImportPath) > 0 {
+		importPaths = append(importPaths, conf.Protobuf.ImportPath)
+	}
+
 	var err error
-	if p.operator, err = strToProtobufOperator(conf.Protobuf.Operator, conf.Protobuf.Message, conf.Protobuf.ImportPath); err != nil {
+	if p.operator, err = strToProtobufOperator(conf.Protobuf.Operator, conf.Protobuf.Message, importPaths); err != nil {
 		return nil, err
 	}
 	return p, nil